@@ -6,12 +6,12 @@ import 'src/mocks.dart';
 
 void main() {
   group('NotesApi for Issues', () {
-    MockGitLabHttpClient mockHttpClient;
+    late MockGitLabHttpClient mockHttpClient;
     GitLab gitLab;
-    IssueNotesApi issueNotes;
+    late IssueNotesApi issueNotes;
 
     final projectId = 1337;
-    final issue = Issue.fromJson(data.decodeMap(data.issue));
+    final issue = Issue.fromJson(data.decodeMap(data.issue)!);
 
     final notesJson = data.decodeList(data.issueNotes);
 
@@ -22,7 +22,7 @@ void main() {
     });
 
     test('Note class properly maps the JSON', () async {
-      for (var item in notesJson) {
+      for (var item in notesJson!) {
         final noteJson = item as Map<String, dynamic>;
         final authorJson = noteJson["author"] as Map<String, dynamic>;
 
@@ -59,7 +59,7 @@ void main() {
       final note = await issueNotes.get(302);
 
       call.verifyCalled(1);
-      expect(note.id, 302);
+      expect(note!.id, 302);
     });
     test('.list()', () async {
       final call = mockHttpClient.configureCall(
@@ -71,7 +71,7 @@ void main() {
 
       call.verifyCalled(1);
       expect(notes, hasLength(2));
-      expect(notes.first.id, 302);
+      expect(notes!.first!.id, 302);
     });
     test('.create()', () async {
       final call = mockHttpClient.configureCall(
@@ -83,7 +83,7 @@ void main() {
       final note = await issueNotes.create("Hello");
 
       call.verifyCalled(1);
-      expect(note.body, "Hello");
+      expect(note!.body, "Hello");
     });
     test('.update()', () async {
       final call = mockHttpClient.configureCall(
@@ -95,7 +95,7 @@ void main() {
       final note = await issueNotes.update(42, "World");
 
       call.verifyCalled(1);
-      expect(note.body, "World");
+      expect(note!.body, "World");
     });
     test('.delete()', () async {
       final call = mockHttpClient.configureCall(
