@@ -6,13 +6,13 @@ import 'src/mocks.dart';
 
 void main() {
   group('MergeRequestsApi', () {
-    MockGitLabHttpClient mockHttpClient;
+    late MockGitLabHttpClient mockHttpClient;
     GitLab gitLab;
-    ProjectsApi project;
+    late ProjectsApi project;
 
     final projectId = 1337;
 
-    final mergeRequestsMap = data.decodeMap(data.mergeRequest);
+    final mergeRequestsMap = data.decodeMap(data.mergeRequest)!;
     final mergeRequestId = mergeRequestsMap['id'] as int;
     final mergeRequestIid = mergeRequestsMap['iid'] as int;
 
@@ -36,7 +36,7 @@ void main() {
       expect(mergeRequest.title, mergeRequestsMap['title']);
       expect(mergeRequest.state, mergeRequestsMap['state']);
 
-      expect(mergeRequest.author.id, mergeRequestsMap['author']['id']);
+      expect(mergeRequest.author!.id, mergeRequestsMap['author']['id']);
       expect(
           mergeRequest.assignees[0].id, mergeRequestsMap['assignees'][0]['id']);
 
@@ -59,7 +59,7 @@ void main() {
       final mergeRequest = await project.mergeRequests.get(mergeRequestId);
 
       call.verifyCalled(1);
-      expect(mergeRequest.id, mergeRequestId);
+      expect(mergeRequest!.id, mergeRequestId);
     });
     test('.list()', () async {
       final call = mockHttpClient.configureCall(
@@ -87,7 +87,7 @@ void main() {
       );
 
       call.verifyCalled(1);
-      expect(mergeRequest.title, "Hello");
+      expect(mergeRequest!.title, "Hello");
     });
     test('.update', () async {
       final call = mockHttpClient.configureCall(
@@ -100,7 +100,7 @@ void main() {
           await project.mergeRequests.update(42, title: "World");
 
       call.verifyCalled(1);
-      expect(mergeRequest.title, "World");
+      expect(mergeRequest!.title, "World");
     });
     test('.update -- change assignees', () async {
       final call = mockHttpClient.configureCall(
