@@ -6,16 +6,16 @@ class IssuesApi {
 
   IssuesApi(this._gitLab, this._project);
 
-  Future<Issue> get(int id) async {
+  Future<Issue?> get(int id) async {
     final uri = _project.buildUri(['issues', '$id']);
 
-    final json = await _gitLab.request(uri) as Map;
+    final json = await _gitLab.request(uri) as Map?;
 
-    return new Issue.fromJson(json);
+    return json == null ? null : Issue.fromJson(json);
   }
 
   Future<List<Issue>> closedByMergeRequest(int mergeRequestIid,
-      {int page, int perPage}) async {
+      {int? page, int? perPage}) async {
     final uri = _project.buildUri(
       ['merge_requests', '$mergeRequestIid', 'closes_issues'],
       page: page,
@@ -28,13 +28,13 @@ class IssuesApi {
   }
 
   Future<List<Issue>> list(
-      {IssueState state,
-      IssueOrderBy orderBy,
-      IssueSort sort,
-      String milestone,
-      List<String> labels,
-      int page,
-      int perPage}) async {
+      {IssueState? state,
+      IssueOrderBy? orderBy,
+      IssueSort? sort,
+      String? milestone,
+      List<String>? labels,
+      int? page,
+      int? perPage}) async {
     final queryParameters = <String, dynamic>{};
 
     if (state != null) queryParameters['state'] = enumToString(state);
@@ -54,18 +54,18 @@ class IssuesApi {
   /// Creates a new issue.
   ///
   /// See https://docs.gitlab.com/ee/api/issues.html#new-issue
-  Future<Issue> add(
+  Future<Issue?> add(
     String title, {
-    String description,
-    bool confidential,
-    List<int> assigneeIds,
-    int milestoneId,
-    List<String> labels,
-    DateTime dueDate,
-    int mergeRequestToResolveDiscussionsOf,
-    String discussionToResolve,
-    int weight,
-    int epicId,
+    String? description,
+    bool? confidential,
+    List<int>? assigneeIds,
+    int? milestoneId,
+    List<String>? labels,
+    DateTime? dueDate,
+    int? mergeRequestToResolveDiscussionsOf,
+    String? discussionToResolve,
+    int? weight,
+    int? epicId,
   }) async {
     final queryParameters = <String, dynamic>{
       "title": title,
@@ -89,27 +89,27 @@ class IssuesApi {
     final json = await _gitLab.request(
       uri,
       method: HttpMethod.post,
-    ) as Map<String, dynamic>;
+    ) as Map<String, dynamic>?;
 
-    return Issue.fromJson(json);
+    return json == null ? null : Issue.fromJson(json);
   }
 
   /// Updates an existing issue.
   ///
   /// See https://docs.gitlab.com/ee/api/issues.html#edit-issue
-  Future<Issue> update(
+  Future<Issue?> update(
     int issueIid, {
-    String title,
-    String description,
-    bool confidential,
-    List<int> assigneeIds,
-    int milestoneId,
-    List<String> labels,
-    DateTime dueDate,
-    int mergeRequestToResolveDiscussionsOf,
-    String discussionToResolve,
-    int weight,
-    int epicId,
+    String? title,
+    String? description,
+    bool? confidential,
+    List<int>? assigneeIds,
+    int? milestoneId,
+    List<String>? labels,
+    DateTime? dueDate,
+    int? mergeRequestToResolveDiscussionsOf,
+    String? discussionToResolve,
+    int? weight,
+    int? epicId,
   }) async {
     final queryParameters = <String, dynamic>{
       if (title != null) "title": title,
@@ -136,9 +136,9 @@ class IssuesApi {
     final json = await _gitLab.request(
       uri,
       method: HttpMethod.put,
-    ) as Map<String, dynamic>;
+    ) as Map<String, dynamic>?;
 
-    return Issue.fromJson(json);
+    return json == null ? null : Issue.fromJson(json);
   }
 
   /// Deletes an existing issue.
@@ -168,17 +168,17 @@ class Issue {
 
   int get iid => originalJson['iid'] as int;
 
-  String get title => originalJson['title'] as String;
+  String get title => originalJson['title'] as String? ?? '';
 
-  String get description => originalJson['description'] as String;
+  String get description => originalJson['description'] as String? ?? '';
 
-  String get state => originalJson['state'] as String;
+  String get state => originalJson['state'] as String? ?? '';
 
   List<String> get labels => (originalJson['labels'] as List).cast<String>();
 
-  String get webUrl => originalJson['web_url'] as String;
+  String get webUrl => originalJson['web_url'] as String? ?? '';
 
-  User get author => originalJson['author'] == null
+  User? get author => originalJson['author'] == null
       ? null
       : User.fromJson(originalJson['author'] as Map<String, dynamic>);
 
@@ -194,17 +194,17 @@ class Issue {
   DateTime get updatedAt =>
       DateTime.parse(originalJson['updated_at'] as String);
 
-  bool get subscribed => originalJson['subscribed'] as bool;
+  bool? get subscribed => originalJson['subscribed'] as bool?;
 
-  int get userNotesCount => originalJson['user_notes_count'] as int;
+  int? get userNotesCount => originalJson['user_notes_count'] as int?;
 
-  DateTime get dueDate => originalJson['due_date'] == null
+  DateTime? get dueDate => originalJson['due_date'] == null
       ? null
       : DateTime.parse(originalJson['due_date'] as String);
 
-  bool get confidential => originalJson['confidential'] as bool;
+  bool? get confidential => originalJson['confidential'] as bool?;
 
-  int get weight => originalJson['weight'] as int;
+  int? get weight => originalJson['weight'] as int?;
 
   @override
   String toString() => 'Issue id#$id iid#$iid ($title)';
