@@ -1,24 +1,26 @@
 extension JsonMap on Map<String, dynamic> {
-  Map<String, dynamic> getJsonMap(String key) =>
+  Map<String, dynamic>? getJsonMap(String key) =>
       this[key] is Map<String, dynamic>
-          ? this[key] as Map<String, dynamic>
+          ? this[key] as Map<String, dynamic>?
           : <String, dynamic>{};
 
-  int getIntOrNull(String key) => getIntOr(key, null);
+  int? getIntOrNull(String key) => this[key] is int ? this[key] as int : null;
   int getIntOr(String key, int defaultValue) =>
-      this[key] is int ? this[key] as int : defaultValue;
+      getIntOrNull(key) ?? defaultValue;
 
-  String getStringOrNull(String key) => getStringOr(key, null);
+  String? getStringOrNull(String key) =>
+      this[key] is String ? this[key] as String : null;
   String getStringOr(String key, String defaultValue) =>
-      this[key] is String ? this[key] as String : defaultValue;
+      getStringOrNull(key) ?? defaultValue;
 
-  bool getBoolOrNull(String key) => getBoolOr(key, null);
+  bool? getBoolOrNull(String key) =>
+      this[key] is bool ? this[key] as bool? : null;
   bool getBoolOr(String key, bool defaultValue) =>
-      this[key] is bool ? this[key] as bool : defaultValue;
+      getBoolOrNull(key) ?? defaultValue;
 
-  DateTime getISODateTimeOrNull(String key) => getISODateTimeOr(key, null);
+  DateTime? getISODateTimeOrNull(String key) => this[key] is String
+      ? DateTime.tryParse(this[key] as String) ?? null
+      : null;
   DateTime getISODateTimeOr(String key, DateTime defaultValue) =>
-      this[key] is String
-          ? DateTime.tryParse(this[key] as String) ?? defaultValue
-          : defaultValue;
+      getISODateTimeOrNull(key) ?? defaultValue;
 }
