@@ -8,15 +8,15 @@ class PipelinesApi {
 
   PipelinesApi(this._gitLab, this._project);
 
-  Future<Pipeline> get(int id) async {
+  Future<Pipeline?> get(int id) async {
     final uri = _project.buildUri(['pipelines', '$id']);
 
-    final json = await _gitLab.request(uri) as Map;
+    final json = await _gitLab.request(uri) as Map?;
 
-    return new Pipeline.fromJson(json);
+    return json == null ? null : Pipeline.fromJson(json);
   }
 
-  Future<List<Pipeline>> list({int page, int perPage}) async {
+  Future<List<Pipeline>> list({int? page, int? perPage}) async {
     final uri = _project.buildUri(['pipelines'], page: page, perPage: perPage);
 
     final jsonList = _responseToList(await _gitLab.request(uri));
@@ -31,9 +31,9 @@ class Pipeline {
   Pipeline.fromJson(this.originalJson);
 
   int get id => originalJson['id'] as int;
-  String get status => originalJson['status'] as String;
-  String get ref => originalJson['ref'] as String;
-  String get sha => originalJson['sha'] as String;
+  String? get status => originalJson['status'] as String?;
+  String? get ref => originalJson['ref'] as String?;
+  String? get sha => originalJson['sha'] as String?;
 
   @override
   String toString() => 'Pipeline id#$id (Ref: $ref)';
